requirement.txt
absl-py==0.13.0
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
astunparse==1.6.3
boto3==1.17.89
botocore==1.20.89
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
distlib==0.3.1
Django==3.2.3
django-crispy-forms==1.11.2
django-filter==2.4.0
django-storages==1.11.1
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-api-core==1.30.0
google-auth==1.32.1
google-auth-oauthlib==0.4.5
google-cloud-core==1.7.1
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.1
idna==2.10
imutils==0.5.4
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
pdfkit==0.6.1
Pillow==8.2.0
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pyPdf==1.13
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.10.1
six==1.15.0
sqlparse==0.4.1
supervisor==4.1.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.2.0
twilio==6.63.1
typing-extensions==3.7.4.3
urllib3==1.26.5
virtualenv==20.4.7
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
xhtml2pdf==0.2.5
